coolkey: Fix integer overflow parsing combined object
Coverity detected an out-of-bounds read in coolkey_process_combined_object
(CID 353967).

This occurred because `object_offset` was declared as a 16-bit `unsigned short`.
If the uncompressed object exceeded 64KB, incrementing the offset caused a
wrap-around. When this wrapped value was used in the bounds check, the addition
failed to trigger an error, allowing the parser to read incorrect offsets and
potentially overrun the buffer.

Changing `object_offset` to `size_t` prevents the wrap-around and safely
evaluates the bounds check.

+++ src/libopensc/card-coolkey.c
@@ -2002,7 +2002,7 @@ coolkey_process_combined_object(sc_card_t *card, coolkey_private_data_t *priv, u
 	unsigned short compressed_offset;
 	unsigned short compressed_length;
 	unsigned short compressed_type;
-	unsigned short object_offset;
+	size_t object_offset;
 	unsigned short object_count;
 	coolkey_decompressed_header_t *decompressed_header;
 	u8 *decompressed_object = NULL;
